FACES-2321 Move the com.liferay.faces.util.render.ContentTypes interface to package com.liferay.faces.util
