+++ portal/src/main/java/com/liferay/faces/portal/component/inputrichtext/internal/InputRichTextRenderer.java
@@ -31,9 +31,9 @@ import javax.faces.context.ResponseWriter;
 import javax.faces.render.FacesRenderer;
 
 import com.liferay.faces.portal.component.inputrichtext.InputRichText;
+import com.liferay.faces.util.ContentTypes;
 import com.liferay.faces.util.component.Styleable;
 import com.liferay.faces.util.lang.StringPool;
-import com.liferay.faces.util.render.ContentTypes;
 import com.liferay.faces.util.render.RendererUtil;
 
 import com.liferay.portal.kernel.util.PropsKeys;
