+++ reslib/src/main/java/com/liferay/faces/reslib/application/internal/ComboResource.java
@@ -30,9 +30,9 @@ import javax.faces.application.Resource;
 import javax.faces.application.ResourceHandler;
 import javax.faces.context.FacesContext;
 
+import com.liferay.faces.util.ContentTypes;
 import com.liferay.faces.util.logging.Logger;
 import com.liferay.faces.util.logging.LoggerFactory;
-import com.liferay.faces.util.render.ContentTypes;
 
 
 /**
