+++ util/src/main/java/com/liferay/faces/util/ContentTypes.java
@@ -11,7 +11,7 @@
  * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
  * details.
  */
-package com.liferay.faces.util.render;
+package com.liferay.faces.util;
 
 /**
  * @author  Brian Wing Shun Chan
